+++ benchmark/python_interface/example.py
@@ -1,4 +1,4 @@
-from benchmarker import fetch_error_model
+from benchmarker import fetch_error_model, Position
 import os
 
 """
@@ -8,18 +8,18 @@ It customizes the error model to that the phenomenological error model in arXiv:
 """
 
 code_distance = 5
-measurement_rounds = 5
+noisy_measurement_rounds = 5
 
 # this parameters only specify the size of the tailored surface code, but no error rate
-configuration = ["--use_rotated_tailored_code"]
+configuration = ["--code_type", "RotatedTailoredCode"]
 
 
 # now fetch the error model with default probability = 0
-error_model = fetch_error_model(code_distance, code_distance, measurement_rounds, p=0, configuration=configuration)
+error_model = fetch_error_model(code_distance, code_distance, noisy_measurement_rounds, p=0, configuration=configuration)
 
 
 # print out a single qubit in the middle to see what field it has, initially all error rate is configured to 0
-middle_qubit = error_model.at(14, 2, 2)
+middle_qubit = error_model.at(Position(6, 3, 3))
 print(middle_qubit)
 
 
@@ -33,12 +33,12 @@ measurement_error_rate = p
 
 
 # update the error model
-for mt in range(measurement_rounds):
-    t = 11 + mt * 6  # add error before each measurement, the first measurement happens at t = 12
-    for i in range(code_distance * 2 - 1):
-        for j in range(code_distance * 2 - 1):
-            qubit = error_model.at(t, i, j)
-            if qubit is not None:  # rotated code will leave some position no qubits
+for mt in range(noisy_measurement_rounds):
+    t = mt * 6  # add error before each measurement, the first measurement happens at t = 6
+    for i in range(error_model.vertical):
+        for j in range(error_model.horizontal):
+            qubit = error_model.at(Position(t, i, j))
+            if qubit is not None and not qubit.is_virtual:  # rotated code will leave some position no qubits
                 if qubit.type == "Data":
 
                     # single-qubit error rate
@@ -89,16 +89,17 @@ for mt in range(measurement_rounds):
 # instead, I developed a visualization tool of any error model
 # unfortunately, due to technical simplicity I didn't implement arbitrary sized visualization, so currently the code_distance and measurement_rounds must be 5
 # (you can still run benchmark for other code size of course, just without visualization)
-if code_distance == 5 and measurement_rounds == 5:
-    error_model.visualize()  # by default use web server: https://qec.wuyue98.cn, you can easily share the link with others
+if code_distance == 5 and noisy_measurement_rounds == 5:
+    # error_model.visualize()  # by default use web server: https://qec.wuyue98.cn, you can easily share the link with others
     # error_model.visualize("http://localhost:8066", None)  # if you decide to start local server using `cargo run --release -- server --port 8066`, uncomment this line
     pass
 else:
     print("[info] visualization is not supported for code distance and measurement rounds other than 5, skipped")
 
+print(middle_qubit)  # check what's after the change
 
 # you can also save this error model to a file, and run benchmark using this error model later
-#    it can be loaded using `cargo run --release -- tool fault_tolerant_benchmark ...... --load_error_model_from_file /path/to/example.json`
+#    it can be loaded using `cargo run --release -- tool benchmark ...... --load_error_model_from_file /path/to/example.json`
 filepath = os.path.join(os.path.dirname(__file__), f"example.json")
 error_model.save(filepath)